Fix for typing issue: config_file_name cannot be None.
Fix UUID issue get_current_user: when using id column in user table is defined as a UUID type, but the value used to filter the results is a string. Solution is to convert the string (token_data.sub) to a UUID. I came across this issue when replacing  Postgresql with SQLite.